Move implementation of DrawCursor
